Fix underscored functions in mod.rs
